src/analyzer/requirements.txt
matplotlib==3.8.0
networkx==3.2.1
numpy==1.26.4
pandas==2.0.3
pydantic==1.10.13
pyvis==0.3.1
PyYAML==5.3.1
Requests==2.31.0
neo4j==5.14.1